Adds gcov_flush as an API call
This is an interim kludge to allow us to flush gcov coverage data onto
disk so that we can measure test coverage. The gcov library will flush
the *.gcda files to disk only when the process is over. In our case,
this is py.test and we need to display coverage information *before*
py.test itself finishes running (but after the tests execute). I tried
exposing __gcov_flush using a small C extension but that doesn't seem to
cut it. The only way is to have the C library being tested itself have
call and then to call that when each test is done.

This "works for now" but will have to be fixed and I'm filing a bug to
track it.

The Makefile adds the -lgcov to link the gcov libaries as does the
pycpslib_build.py file. The pslib_linux.c file has the API call which
will be called the py.test infrastructure.